Fixed premissions for cloud connections connect disconnect button
