+++ kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/server/GwtCloudConnectionServiceImpl.java
@@ -22,6 +22,7 @@ import static org.eclipse.kura.web.shared.model.GwtCloudConnectionEntry.GwtCloud
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
@@ -33,6 +34,7 @@ import java.util.stream.Collectors;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
+import org.eclipse.kura.KuraConnectException;
 import org.eclipse.kura.KuraException;
 import org.eclipse.kura.cloud.CloudService;
 import org.eclipse.kura.cloud.factory.CloudServiceFactory;
@@ -43,6 +45,7 @@ import org.eclipse.kura.cloudconnection.factory.CloudConnectionFactory;
 import org.eclipse.kura.cloudconnection.publisher.CloudPublisher;
 import org.eclipse.kura.cloudconnection.subscriber.CloudSubscriber;
 import org.eclipse.kura.configuration.ConfigurationService;
+import org.eclipse.kura.data.DataService;
 import org.eclipse.kura.web.server.util.GwtComponentServiceInternal;
 import org.eclipse.kura.web.server.util.ServiceLocator;
 import org.eclipse.kura.web.server.util.ServiceLocator.ServiceConsumer;
@@ -64,6 +67,7 @@ import org.osgi.framework.FrameworkUtil;
 import org.osgi.framework.InvalidSyntaxException;
 import org.osgi.framework.ServiceReference;
 import org.osgi.service.cm.ConfigurationAdmin;
+import org.osgi.service.component.ComponentConstants;
 import org.osgi.service.component.runtime.ServiceComponentRuntime;
 import org.osgi.service.component.runtime.dto.ComponentDescriptionDTO;
 import org.slf4j.Logger;
@@ -81,7 +85,11 @@ public class GwtCloudConnectionServiceImpl extends OsgiRemoteServiceServlet impl
     private static final String CLOUD_PUBLISHER = CloudPublisher.class.getName();
     private static final String CLOUD_SUBSCRIBER = CloudSubscriber.class.getName();
 
+    private static final String DATA_SERVICE_REFERENCE_NAME = "DataService";
+
     private static final Logger auditLogger = LoggerFactory.getLogger("AuditLogger");
+
+    private static final Logger logger = LoggerFactory.getLogger(GwtCertificatesServiceImpl.class);
 
     @Override
     public List<GwtCloudEntry> findCloudEntries() throws GwtKuraException {
@@ -498,5 +506,189 @@ public class GwtCloudConnectionServiceImpl extends OsgiRemoteServiceServlet impl
         final HttpSession session = request.getSession(false);
 
         GwtComponentServiceInternal.updateComponentConfiguration(session, component);
+    }
+
+    @Override
+    public void connectDataService(GwtXSRFToken xsrfToken, String connectionId) throws GwtKuraException {
+        checkXSRFToken(xsrfToken);
+
+        final HttpServletRequest request = getThreadLocalRequest();
+        final HttpSession session = request.getSession(false);
+
+        Collection<ServiceReference<CloudService>> cloudServiceReferences = ServiceLocator.getInstance()
+                .getServiceReferences(CloudService.class, null);
+
+        for (ServiceReference<CloudService> cloudServiceReference : cloudServiceReferences) {
+            String cloudServicePid = (String) cloudServiceReference.getProperty(KURA_SERVICE_PID);
+            if (cloudServicePid.endsWith(connectionId)) {
+                String dataServiceRef = (String) cloudServiceReference
+                        .getProperty(DATA_SERVICE_REFERENCE_NAME + ComponentConstants.REFERENCE_TARGET_SUFFIX);
+                Collection<ServiceReference<DataService>> dataServiceReferences = ServiceLocator.getInstance()
+                        .getServiceReferences(DataService.class, dataServiceRef);
+
+                for (ServiceReference<DataService> dataServiceReference : dataServiceReferences) {
+                    DataService dataService = ServiceLocator.getInstance().getService(dataServiceReference);
+                    if (dataService != null) {
+                        GwtKuraException gwtKuraException = null;
+                        int counter = 10;
+                        try {
+                            dataService.connect();
+                            while (!dataService.isConnected() && counter > 0) {
+                                Thread.sleep(1000);
+                                counter--;
+                            }
+                        } catch (KuraConnectException e) {
+                            logger.warn("Error connecting");
+                            gwtKuraException = new GwtKuraException(GwtKuraErrorCode.CONNECTION_FAILURE, e,
+                                    "Error connecting. Please review your configuration.");
+                        } catch (InterruptedException e) {
+                            logger.warn("Interrupt Exception");
+                            gwtKuraException = new GwtKuraException(GwtKuraErrorCode.INTERNAL_ERROR, e,
+                                    "Interrupt Exception");
+                        } catch (IllegalStateException e) {
+                            logger.warn("Illegal client state");
+                            gwtKuraException = new GwtKuraException(GwtKuraErrorCode.INTERNAL_ERROR, e,
+                                    "Illegal client state");
+                        }
+
+                        if (gwtKuraException != null) {
+                            auditLogger.warn(
+                                    "UI CloudConnection - Failure - Failed to connect data service for user: {}, session: {}, connection id: {}",
+                                    session.getAttribute(Attributes.AUTORIZED_USER.getValue()), session.getId(),
+                                    connectionId);
+                            throw gwtKuraException;
+                        }
+                    }
+                    ServiceLocator.getInstance().ungetService(dataServiceReference);
+                }
+            }
+            ServiceLocator.getInstance().ungetService(cloudServiceReference);
+        }
+
+        Collection<ServiceReference<CloudConnectionManager>> cloudConnectionManagerReferences = ServiceLocator
+                .getInstance().getServiceReferences(CloudConnectionManager.class, null);
+
+        for (ServiceReference<CloudConnectionManager> cloudConnectionManagerReference : cloudConnectionManagerReferences) {
+            String cloudConnectionManagerPid = (String) cloudConnectionManagerReference.getProperty(KURA_SERVICE_PID);
+            if (cloudConnectionManagerPid.endsWith(connectionId)) {
+                CloudConnectionManager cloudConnectionManager = ServiceLocator.getInstance()
+                        .getService(cloudConnectionManagerReference);
+                try {
+                    cloudConnectionManager.connect();
+                } catch (KuraException e) {
+                    logger.warn("Error connecting");
+                    auditLogger.warn(
+                            "UI CloudConnection - Failure - Failed to connect data service for user: {}, session: {}, connection id: {}",
+                            session.getAttribute(Attributes.AUTORIZED_USER.getValue()), session.getId(), connectionId);
+                    throw new GwtKuraException(GwtKuraErrorCode.CONNECTION_FAILURE, e,
+                            "Error connecting. Please review your configuration.");
+                }
+            }
+            ServiceLocator.getInstance().ungetService(cloudConnectionManagerReference);
+        }
+
+        auditLogger.info(
+                "UI CloudConnection - Success - Successfully connected data service for user: {}, session: {}, connection id: {}",
+                session.getAttribute(Attributes.AUTORIZED_USER.getValue()), session.getId(), connectionId);
+    }
+
+    @Override
+    public void disconnectDataService(GwtXSRFToken xsrfToken, String connectionId) throws GwtKuraException {
+        checkXSRFToken(xsrfToken);
+
+        final HttpServletRequest request = getThreadLocalRequest();
+        final HttpSession session = request.getSession(false);
+
+        Collection<ServiceReference<CloudService>> cloudServiceReferences = ServiceLocator.getInstance()
+                .getServiceReferences(CloudService.class, null);
+
+        for (ServiceReference<CloudService> cloudServiceReference : cloudServiceReferences) {
+            String cloudServicePid = (String) cloudServiceReference.getProperty(KURA_SERVICE_PID);
+            if (cloudServicePid.endsWith(connectionId)) {
+                String dataServiceRef = (String) cloudServiceReference
+                        .getProperty(DATA_SERVICE_REFERENCE_NAME + ComponentConstants.REFERENCE_TARGET_SUFFIX);
+                Collection<ServiceReference<DataService>> dataServiceReferences = ServiceLocator.getInstance()
+                        .getServiceReferences(DataService.class, dataServiceRef);
+
+                for (ServiceReference<DataService> dataServiceReference : dataServiceReferences) {
+                    DataService dataService = ServiceLocator.getInstance().getService(dataServiceReference);
+                    if (dataService != null) {
+                        dataService.disconnect(10);
+                    }
+                    ServiceLocator.getInstance().ungetService(dataServiceReference);
+                }
+            }
+            ServiceLocator.getInstance().ungetService(cloudServiceReference);
+        }
+
+        Collection<ServiceReference<CloudConnectionManager>> cloudConnectionManagerReferences = ServiceLocator
+                .getInstance().getServiceReferences(CloudConnectionManager.class, null);
+
+        for (ServiceReference<CloudConnectionManager> cloudConnectionManagerReference : cloudConnectionManagerReferences) {
+            String cloudConnectionManagerPid = (String) cloudConnectionManagerReference.getProperty(KURA_SERVICE_PID);
+            if (cloudConnectionManagerPid.endsWith(connectionId)) {
+                CloudConnectionManager cloudConnectionManager = ServiceLocator.getInstance()
+                        .getService(cloudConnectionManagerReference);
+                try {
+                    cloudConnectionManager.disconnect();
+                } catch (KuraException e) {
+                    logger.warn("Error disconnecting");
+                    auditLogger.warn(
+                            "UI CloudConnection - Failure - Failed to disconnect data service for user: {}, session: {}, connection id: {}",
+                            session.getAttribute(Attributes.AUTORIZED_USER.getValue()), session.getId(), connectionId);
+                    throw new GwtKuraException(GwtKuraErrorCode.CONNECTION_FAILURE, e,
+                            "Error disconnecting. Please review your configuration.");
+                }
+            }
+            ServiceLocator.getInstance().ungetService(cloudConnectionManagerReference);
+        }
+
+        auditLogger.info(
+                "UI CloudConnection - Success - Successfully disconnected data service for user: {}, session: {}, connection id: {}",
+                session.getAttribute(Attributes.AUTORIZED_USER.getValue()), session.getId(), connectionId);
+    }
+
+    @Override
+    public boolean isConnected(GwtXSRFToken xsrfToken, String connectionId) throws GwtKuraException {
+        checkXSRFToken(xsrfToken);
+        boolean isConnected = false;
+
+        Collection<ServiceReference<CloudService>> cloudServiceReferences = ServiceLocator.getInstance()
+                .getServiceReferences(CloudService.class, null);
+
+        for (ServiceReference<CloudService> cloudServiceReference : cloudServiceReferences) {
+            String cloudServicePid = (String) cloudServiceReference.getProperty(KURA_SERVICE_PID);
+            if (cloudServicePid.endsWith(connectionId)) {
+                String dataServiceRef = (String) cloudServiceReference
+                        .getProperty(DATA_SERVICE_REFERENCE_NAME + ComponentConstants.REFERENCE_TARGET_SUFFIX);
+                Collection<ServiceReference<DataService>> dataServiceReferences = ServiceLocator.getInstance()
+                        .getServiceReferences(DataService.class, dataServiceRef);
+
+                for (ServiceReference<DataService> dataServiceReference : dataServiceReferences) {
+                    DataService dataService = ServiceLocator.getInstance().getService(dataServiceReference);
+                    if (dataService != null) {
+                        isConnected = dataService.isConnected();
+                    }
+                    ServiceLocator.getInstance().ungetService(dataServiceReference);
+                }
+            }
+            ServiceLocator.getInstance().ungetService(cloudServiceReference);
+        }
+
+        Collection<ServiceReference<CloudConnectionManager>> cloudConnectionManagerReferences = ServiceLocator
+                .getInstance().getServiceReferences(CloudConnectionManager.class, null);
+
+        for (ServiceReference<CloudConnectionManager> cloudConnectionManagerReference : cloudConnectionManagerReferences) {
+            String cloudConnectionManagerPid = (String) cloudConnectionManagerReference.getProperty(KURA_SERVICE_PID);
+            if (cloudConnectionManagerPid.endsWith(connectionId)) {
+                CloudConnectionManager cloudConnectionManager = ServiceLocator.getInstance()
+                        .getService(cloudConnectionManagerReference);
+
+                isConnected = cloudConnectionManager.isConnected();
+            }
+            ServiceLocator.getInstance().ungetService(cloudConnectionManagerReference);
+        }
+
+        return isConnected;
     }
 }
