+++ kura/org.eclipse.kura.web2/src/main/java/org/eclipse/kura/web/shared/service/GwtCloudConnectionService.java
@@ -133,4 +133,24 @@ public interface GwtCloudConnectionService extends RemoteService {
 
     public void updateStackComponentConfiguration(GwtXSRFToken xsrfToken, GwtConfigComponent component)
             throws GwtKuraException;
+
+    /**
+     * Connects the local MQTT client to the specified broker.
+     *
+     * @param xsrfToken
+     *            - A GwtXSRFToken token necessary to prevent cross-site request forgery attacks.
+     * @throws GwtKuraException
+     */
+    public void connectDataService(GwtXSRFToken xsrfToken, String connectionId) throws GwtKuraException;
+
+    /**
+     * Disconnects the local MQTT client.
+     *
+     * @param xsrfToken
+     *            - A GwtXSRFToken token necessary to prevent cross-site request forgery attacks.
+     * @throws GwtKuraException
+     */
+    public void disconnectDataService(GwtXSRFToken xsrfToken, String connectionId) throws GwtKuraException;
+
+    public boolean isConnected(GwtXSRFToken xsrfToken, String connectionId) throws GwtKuraException;
 }
